+++ src/logic/likes_logic.py
@@ -7,6 +7,21 @@ class LikesLogic:
  
     #👍 user like a vacation
     def add_like(self, userID, vacationID):
+
+        #check if userID exists:
+        check_user = "SELECT * FROM vacation.users WHERE userID = %s"
+        user_params = (userID, )
+        result = self.dal.get_scalar(check_user, user_params)
+        if not result:
+            return "User Doesn't Exist. Please try again."
+        
+        #check if vacationID exists:
+        check_vacation = "SELECT * FROM vacation.vacations WHERE vacationID = %s"
+        user_params = (vacationID, )
+        result = self.dal.get_scalar(check_vacation, user_params)
+        if not result:
+            return "Vacation Doesn't Exist. Please try again."
+
         #check if like already exists:
         check_sql = """
                 SELECT count(*) FROM vacation.likes
@@ -28,12 +43,29 @@ class LikesLogic:
                 return f"Last row ID: {result}\nuserID {userID} liked vacationID {vacationID}"
             else:
                 return f"No likes added."
-        
-        else:
+            
+        elif check_result[0]['count(*)'] != 0:
             return "User already liked this vacation."
+        
+
 
     #👍 user unlike vacation
     def delete_like(self, userID, vacationID):
+
+        #check if userID exists:
+        check_user = "SELECT * FROM vacation.users WHERE userID = %s"
+        user_params = (userID, )
+        result = self.dal.get_scalar(check_user, user_params)
+        if not result:
+            return "User Doesn't Exist. Please try again."
+        
+        #check if vacationID exists:
+        check_user = "SELECT * FROM vacation.vacations WHERE vacationID = %s"
+        user_params = (vacationID, )
+        result = self.dal.get_scalar(check_user, user_params)
+        if not result:
+            return "Vacation Doesn't Exist. Please try again."
+        
         sql = """
         DELETE FROM vacation.likes
         WHERE userID = %s AND  vacationID = %s;
